Edit reducers. Edit folder structure. Delete input field in BookForm.
