+++ src/components/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import BooksList from './BooksList';
-import BooksForm from './BooksForm';
+import BooksList from '../containers/BooksList';
+import BooksForm from '../containers/BooksForm';
 
 export default function App() {
   return (
