+++ src/reducers/index.js
@@ -1,34 +1,6 @@
-const mockUpBooks = [
-  {
-    id: Math.floor(Math.random() * 100),
-    title: 'The old man and the sea',
-    category: 'Kids',
+import { combineReducers } from 'redux';
+import booksReducer from './books';
 
-  },
-  {
-    id: Math.floor(Math.random() * 100),
-    title: 'The knight in the rusty armor',
-    category: 'Action',
-
-  },
-  {
-    id: Math.floor(Math.random() * 100),
-    title: 'Alive',
-    category: 'Action',
-
-  },
-  {
-    id: Math.floor(Math.random() * 100),
-    title: 'The Butterfly Effect',
-    category: 'Sci-Fi',
-
-  },
-  {
-    id: Math.floor(Math.random() * 100),
-    title: 'Secret weapons of World War II',
-    category: 'History',
-
-  },
-];
-
-export default mockUpBooks;
+export default combineReducers({
+  booksReducer,
+});
